Add capability to set application constraints
In the same vein as work previously done to enable the setting of model
constraints (see issue #109 / PR #118), this commit introduces the
capability to set constraints on individual applications.

This requires the addition of a read-only attribute, "principal". This
attribute reflects whether the application is principal or not (i.e.
subordinate), which is used to avoid errors when trying to get or set
constraints for subordinates.

As this is a read-only value which is depended on by a user-configurable
value, the Create and Update functions have been updated to trigger a
Read action immediately after running, ensuring that all values are
accurate.