Check and fail if constraints or series is given with manual provisioning
So the scenario is that we provide the ssh_address of a machine that
we want to manually provision, along with some `constraints`
(e.g. mem=8G) or `series`. Normally this is something that juju cli
ignores. However, based on discussions with @tiradojm, we'd like avoid
having the user under any assumptions that these constraints and
series will be used in the addressed machine, so we error instead of
ignoring input.