Use ConflictsWith to make series/constraints/ssh_address mutually exclusive
One caveat of making these mutually exclusive is that we no longer can
make the `series` a required field. However, we do want it to be
required unless the `ssh_address` is provided. So we manually error if
that's the case, because there's no good way of expressing that with
the schema.