No need to load the Juju client keys
Since we have the public_key_file and the private_key_file required
arguments, so the provider doesn't need to be able to auto-detect
juju's own keys (like the juju cli does). This is also good because we
don't need to import as much packages from Juju into the provider.