Use modelInfo to get the default series if not set by argument
The series argument is used to be a required argument. We made it
optional because it conflicts with the manual provision arguments
(e.g. ssh_address). Initially we set it up to error if the series is
not provided, but now we're getting the default series from the
modelInfo like the juju/cmd (2.9) suggests. This will probably change
when we move to 3.x as a part of the series->base effort.