Improve Makefile and README.md.
First attempt at developing in the juju terraform provider
repo highlighted a improvements to make. Added new make
targets to build, lint and copy the provider to the appropriate
directory for use. Updated the README will new make targets.

Attempt to add envtestlxd target.

All of the env var works but JUJU_CA_CERT. Not sure what is going
on, but the provider fails to recognize the result as a cert. More
to do later.