Add new terraform framework juju provider.
The terraform sdk2 has been replaced by the terraform framework. The
new juju provider will replace the current sdk provider, shortly. Until
all of the data sources and resources are transfered to the framework,
both style of provider is required.

No changes to current functionality are made during this transition.