Create a mux to use the terraform sdk2 and framework together.
During the transition from using the terraform sdk2 to the terraform
framework, use a mux as a transition point. This transition temporarily
will required 2 providers, one of each style. Data sources and resources
can be moved from old to new styles as they are completed. When done,
the mux and sdk2 provider will be removed.