Replace sdk model data source with framework version.
Transform the current sdk data source into one using the terraform
plugin framework. Data source data is now a structure and easier to
reason about. ID is required in the structure only for use by the
testing framework.

A new acceptance test was added to handle both the sdk and framework
versions of the juju provider. Our AT have overlapping concerns with
other resources and data sources which make tests a bit difficult while
the mux is in place. They can be cleaned up once the mux has been
removed. The origin test case now runs against the released 0.8.0
version of the provider.