Collection of review changes.
Added todo for name and displayname to be updated, the current
implementation violates terraform's declarative model.

Check r.client before using to prevent panics. This is a paranoia check,
as if it happens, there is most likely a programming error.

Use the userNameFromID method, instead of creating the ID uniquely in
places.

Verify changes are needed before making any changes in Update.