Accommodate the upgade coming from buggy access model resource 0.8.0
In 0.8.0 sdk2 version of the provider, the implementation of the
access model resource had a bug where it didn't contain the users. So
we accommodate upgrades from that by attempting to get the users from
the state, if the ID doesn't contain any users (which happens only
when coming from the previous version because the ID is a computed
field).