Migrate the application resource from the sdk to framework style.
There are no changes to the schema.

Introduce retry read application on not found error.

With the sdk, we could call the readApplication method from
createApplication method to fill in data only known after the
application is deploy such as is principal and placement. With the
framework this is not directly possible. Received application not found
errors.

The work around of setting values until the next read causes tests to
fail. Introduced a ReadApplicationWithRetryOnNotFound method to ensure
we get the application data if we know the application exists.

Make a map before using, avoid assignment to nil map panics.

Could happen if config is added, when it wasn't there before. Usually
appeared on k8s, not machines.