Change config to map[string]string.
With the TF framework, there is no longer a need to handle config as a
map[string]interface and do many convertions. Integers, strings and
booleans can all be added to the plan as strings, then converted by the
juju controller based on the actual config type defined by the charm.