Use model name with client's GetConnection.
Create a cache of model data in the ConnectionFactory for use by
GetConnection. Model data saved is the UUID, Type and Name, all of which
are immutable once the model is created. GetConnection can then be
called with a model name, which is easily available in the terraform plans.

It's possible to fill the model cache at once when subsequent terraform
operations are preformed after a model is created as there is no juju
facade to get model details by name.

This does 2 things, reduces the juju api calls necessary, and removes
the burden of finding the model UUID and Type from the provider, which
has no need of the info.