Fix tests and cleanup for modelResource
Address review comments
Remove oldjuju for juju_model resources in tests
Take out the PlanOnly from plan checker for the test
Only write cloud and constraints, if they are set in the plan.
Otherwise the next terraform plan will think there is a change, when
there is not.
Ensure we use a []nestedCloud getting the Cloud from state.
Fix ImportState and cleanup PlanCheck
Add changes to internal structs for model
Fix linter
Address review & fix ImportState passthrough with modified ID
Check client in ImportState for modelResource
Clear out the oldjuju in tests for model resources
Fix import with cloud.

Cloud is a NestedBlock which cannot be computed. If we write it to state
when not specified in plan, the next call to `terraform plan` will try
to recreate it. However we must write something at import. Usually it
will not be set. So add a key in the private data to indicate whether
the model has been imported. If it has, write the cloud info. Otherwise
anly write if it's been written before, though a model's cloud is
immutiable on the juju side.

With import, it's up to the user to ensure that all required data is in
the plan. Most documents suggest reading the terraform state data and
updating the imported resource to match.