Clean migration code in tests
This does two things: 1) keeps the migration test as _Edge tests along
with the Stable tests (which download and use the upstream stable
release of the proider). 2) During the migration, we had to duplicate
the plans used in tests to be able to have some resources be provided
by the "oldjuju". This removes those duplicate plans, and have both
the Edge and Stable tests run the same configration.