Fix remove integration with multiple consumers
Currently removing an integration with an offer removes
the remote offer. This does not take into consideration
of any other consumers integrated with the offer.
This leads to other consumers getting relation-broken
which is not expected.

Remove the integration using DestroyIntegration irrespective
of whether the endpoint is an offer or not.

The acceptance tests that involves postgres deployment takes
around 5 minutes and the new test added also requires
deploying postgres which increases the total time required
for integration tests to run. So increase integration tests
timeout to 40m from 30m.

Fixes: #308