Recreate CreateModelResponse as a stand alone object.
Juju implementation details are leaking from the internal juju library
into the provisioner code. E.G. the provisioner does not need to now the
difference between a credential name and tag. This also reduces package
dependencies. This is one step along the way to separation of concerns.

Having the provision handle terraform and juju details, led to issue
310, where two things called CloudCredential were not equivalent.