Handle ConsumeRemoteOffer when saas application already exists
Creation of multiple resource integrations to same offer may lead
to error for one of the integration as both the integrations tries
to add remote offer to the model. One of them errors out with message
saas application already exists.
Check for the error message while consuming the remote offer. If the
error message says saas application is already created, proceed with
the integration instead of erroring out.