Fix update application related to charm piece.
Remove support for changing the operating system during an update phase.
The way this was done could inadvertently install a charm with the wrong
operating system on existing units during a refresh.

Fix updating the charm revision and channel. This must be done with the
same operating system and architecture as the original charm was
deployed with.

Move changing application constraints to before