Anticipate LP 2039179.
In juju 2.9.45, ResolveCharms may return a charm and sugested operating
system which do not reflect or support the operating system requested by
the user. Given an appropriate error to the user if so, rather than
continuing to deploy a charm unsupported by the user specified operating
system.

Update UT to use charms and charm/operating system combinations which
will not hit this bug.