Add Base to application resources.
Will replace series which is now deprecated. Due to use of the 2.9.45
code base, most operating system pieces will be handled as a series
still. But this gives users a transition to be able to use bases.

Remove unused elemented of CreateApplicationResponse. ReadApplication
is necessary for all required data anyway.