Deprecate Principal in Application schema.
It's a value only used by the terraform provider and not the users.
Nor does the provide use the value saved, rather it uses the value
from a juju api call.

Leave the element in the application data structure to allow upgrades
of the provider without versionining the schema. That will be done for
version 1.0.0