Add lifecycle hint to integration example.
While testing update of an application resource with integrations where
RequiresReplace is hit, found an issue with integrations. Removal of an
application, also removes any integrations used by the application. This
causes state confusion for terraform. Adding the replaced_triggered_by
lifecycle info, allows for the integration to also be replaced if the
related application resources are.

List specific schema attributes so that no any change to the application
will trigger the integration as well.