A better way to find the controller version.
The old way failed when a juju user without permissions for the
controller model used a plan to deploy an application. It also
created an additional connection and client, which with the new
method are not needed.

Leaving the controller version off the application client is not a
problem as a new client is created for all terraform plans/applies.
It will be found new each time.