Ensure we have a suggested series from resolveCharm.
Between juju 2.9 and 3.x, there is a move from using series to base.
This provider client is using the juju 2.9.46 code base for API, but may
be run against 2.9, 3.1.x, 3.3.x controller with the expection that
current functionality will work. Trust the base first as 3.1 may not
return a series in the charm.URL. Fall back is the series from the URL.