Handle resolved charm origins without base.
In the case of resolving a charm with a revision, unless the user supplies a base to use, ResolveCharm will not return a suggested base. However there will be a suggested list of series or bases which can be used with that charm revision. Log any errors from GetSeriesFromBase for using an empty base and continue rather than failing. seriesToUse will handle picking a base.

When juju queries charmhub about a charm by default or with a specific channel, the base negotiated starting with the architecture. If you specify a revision, charmhub assumes you know the base or at least the architecture and it's not part of the request.