Fix errors found by update to terraform-plugin-framework v1.5.0.
Get the actual machine base and series data after it's been started.
Cannot assume that the plan has any or all of the data as we had been
doing. Added a read machine with retry method to wait for the data.

Fix where the manual machine provisioning was not returning the
machine base.