Use go-version-file with actions/setup-go.
Some of the places where setup-go was used, didn't specify a go version
or where to find the version. After the update of go.mod to use go 1.21,
some of the workflows started to fail due to the version of go. Avoid
this in the future by adding go-version-file to the places it's missing.