Test with the juju 3/stable snap channel.
The 3 track for juju snaps is the recommended version of 3 to be using and
will continue to update so we don't have to update the minor versions that
frequently.

Juju 3.4.0 will be released before 3.3.2. Either will have the fix we need
to fix the panic for not verifying a map entry before using.