Implement endpoint bindings for applications (#400) (#418)
* Implement endpoint bindings for applications (#400)

Add the attributes `endpoint_bindings` to the application resource. This
endpoint is a set nested attribute.

Usage:
```tf
endpoint_bindings = [
  {
    "space"    = "my_space1",
  },
  {
    "endpoint" = "my_endpoint"
    "space"    = "my_space2"
  }
]
```

* Handle import state better for endpoint bindings

Add notice to remove validation in client side until bug lp#2055868 is
resolved.

Use model's default space to compare what's a default or not from an
application.

* Simplify default space handling on read app

Make sure case where default space is not defined, alpha is used as
default space.

* Bind endpoints to default space when endpoint_bindings is null

When endpoint_bindings definition is removed from an application,
interpret it as intent and go back to default space.