Save the secret URI ID not the schema.
When interacting with secrets, most do not know there is a schema
associated with it. To ensure we don't get mismatches between the two
types with terraform doing a string comparision, only save the ID. The
risk is with the Access Secret resource where we cannot guarentee how
the secret id is provided. Related docs will be updated to be more
clear.