feat(resources): address review comments to update to latest resource revision
upon channel updates

If the charm channel is updated and revision or custom image is not
specified in the TF plan, a resource revision should be attached
according the updated channel.
Formerly, this functionality is not implemented properly.
This commit implements this.

revert: Unncessary resource files added for testing are removed.

chore: Some libraries are updated in go.mod and go.sum.
Linting issues are fixed.

docs(resources): Documentation is updated for `resources` key
to include added functionality.

Signed-off-by: gatici <gulsum.atici@canonical.com>