feat: add jaas validators
The AvoidJAASValidator can be used to ensure a resource is not used with JAAS.
E.g. juju_access_model should not be used with JAAS as there is a JAAS specific type for that.

The RequiresJAASValidator can be used to ensure a resource is only used with JAAS.
E.g. new JAAS specific resources should not be used when communicating with a Juju controller.