chore(constraints): constraints are not computed
Resolves: https://github.com/juju/terraform-provider-juju/issues/466

Juju does not distinguish between null constraints and empty string
constraints. However, terraform does not consider null to be a computed
value, but it does the empty string.

This means that on the occasion we end up with a null value for
constraint (such as the import of a subordinate application), on next
apply we will erroneously attempt to compute the constraints by setting
them.

This leads to failures when updating imported subordinate applications,
since we cannot set constraints for subordinates.

Removing the computed flag from constraints informs terraform that null
valued constraints are valid and do not need to be computed.