feat: add computed uuid to model resource
Add uuid computed attribute to the juju_model resource. It already exists in the
juju_model datasource.

The sharedClient ModelUUID method will return the uuid if that's what
provided as the ModelIdentifier, otherwise it will check the model cache
to convert the model name to a model uuid.

The model cache has been updated to key off the model uuid rather than
name to faciliate differentiating models of the same name with different
users.