feat(update-app-base): add support to update `base` for application charms
This PR adds support to update the base in application charms by requiring a replace in case of a
machine charm, and perform the upgrade in case of a k8s charm.

We add the model_type computed field on the application schema, and we
use it to make a decision in the planmodifier RequiresReplaceIf.

re #635