fix: fix the use of a hardcoded admin user
The terraform provider code used a hardcoded `admin` user in a number of place, which caused issues
if a different username (or client credentails) was used. With this change we use the username (or
clientID) as specified in the provider.

This PR also changes the username that is passed in arguments to SSH key management calls. Juju
3.6 does not associate ssh keys with users - ssh keys are global per model - so the username
is ignored. For consistency we pass in the username that is used to log in to the
controller because in Juju 4 ssh keys will be associated with users.