fix: test TestAcc_ResourceApplication_UpdatesRevision
We set the ID and Hash as an empty value to trick juju into installing
the right version of the revision specified. The right empty values was
already set in newOrigin.ID and newOrigin.Hash but not copied to
oldOrigin, which is the one used to resolve charm to be deployed.