feat(applications): use custom oci images from the state
In this PR we are using OCI images from the state rather then relying on the revision number coming
from the juju client to avoid the bug reported at #663. To do so, we are special casing resources
with the origin set to OriginUpload. By setting the revision to -1, we are able later one to ignore
the value coming from the resource client but rather using the value we have in the state.