feat(applications): track custom oci images in the resources
In this PR we are tracking OCI images in the state to avoid the bug reported at #663. To do so, we
are special casing resources with the origin set to OriginUpload. By setting the revision to -1, we
are able later one to ignore the value coming from the resource client but rather using the value we
have in the state.