feat(wait-for functions): add two new generic methods called `WaitFor` and `WaitForError`
These two methods let you specify the condition by which you want to wait for getting info from the
juju client. The can be useful to wait for updates to be propagated, or to wait for specific error
(ex. ApplicationNotFound to make sure the application has been deleted after Destroy has been
called).