fix(test): fix TestAcc_CustomResourcesAddedToPlanMicrok8s by time.Sleep
There is a Juju timing issue by which if the app is created and after it the resource is updated
sometimes the update is not applied. This is most likely an issue that is not affecting any user
because it is just a scenario presented in our tests.The fix is just to sleep between step 1 and
step 2.