refactor: enable default model arch to be specified via env var
Enables deriving the architecture from an envionrment variable called
JUJU_DEFAULT_TEST_MODEL_ARCHITECTURE. This was required because juju defaults the architecture of
workload machines to amd64 unless otherwise specified. When running the TF acceptance tests on
anything other than amd64, it would fail because the constraints would not match on application
deploy. To fix this, we now have a small constraint injection at the client level when creating
models. A side piece of this commit is cleaning up parameters to be passed to the client, should we
ever need more. This prevents us bloating the NewClient signature.