chore: store machine ID after creation
Recent changes to the provider ensure that we wait for provisioned machines to become available
before continuing.
However, if we timeout while waiting or there is some kind of networking issue between the
provider and Juju controller, we lose track of the machine we created.
This change saves the ID of the created resources after creation rather than saving it after
waiting.
The benefit of this is that when the plan is reapplied, the Read method will trigger because
Terraform understands a machine already exists.

To support the above, I've moved the logic to wait-for a machine's readiness into a separate
routine so that it can be called from Read and Create.

Finally, it is unfortunately difficult to test this scenario without unit tests that can induce
a "failure" in the wait-for logic.