fix(delete-multiple-integrations-same-endpoint): fix delete multiple integrations with same endpoint
Before the deletion was failing because the integration <provider:endpoint> <requirer:endpoint> was
deleted using just the <requirer:endpoint> part. In case we have multiple equal <requirer:endpoint>
as in the test added in this pr, Juju can't disambiguate which it has to delete. In this PR the
deletion is done by using the fully qualified integration endpoints string.