fix(update-offer-access): refactor code for update offer access
Before the logic was to revoke all of the users' accesses and re-granting it to level set in the
plan. This can cause offers to reach a "suspended" status. Now the code evaluates if the access
needs to be revoked and granted to an higher level. I've preserved the switch statemenet with all
use cases, even the passthrough, to improve clarity.