feat: make fetching an offer's model UUID optional
The offer data source contains a modelName field that was populated based on the offer URL.
An existing TODO "Is model necessary at all?" highlights that the modelname is unnecessary
because it goes unused and isn't needed for consuming the offer.
With the move to model UUIDs things get trickier because although a user has access to an offer,
they may not have access to the model where the offer is hosted.
To workaround this I've added a flag to the GetOffer method
to make it optional to fetch the model UUID for an offer.