test: remove microk8s setup for jaas test and refactor tests
In this commit we get rid of the microk8s setup for JAAS test because it
was an overhead which wasn't testing anything in the JAAS scenario. I've
also refactored the k8s tests because the test's template was overloaded
with behaviors to support both jaas and not jaas. I've resorted to
create two different plan. And I've also removed the workaround to avoid
the delete because now that we wait for model's deletion we can delete
clouds as well.