feat: retry on any error during resource destruction
Before this change, certain tests were failing with the error,
"could not filter units: could not filter units: cannot get status: agent not found (not found)".

This is happening because the `Status()` call to Juju was failing.
During destroy, repeat the read call on any API error until we get the expected error or timeout.