test: deploy app in k8s on lxd tests
The K8s on LXD tests sometimes fail when there is high concurrency on tests with the error "Unable to remove kubernetes cloud, got error removing kubernetes cloud: cloud is used by 1 model".
This is likely because the model is created and destroyed too quickly based on what we've seen in other tests. To try and resolve this we can also deploy an app.