fix: fix manual provisioning
The implementation for manual provisioning was off, in that it didn't respect ubuntu users could
pre-exist and presumed users will ahve ~/.ssh populated (the same presumption as the cli). This
isn't always the case and probably isn't for a TF user. We now provide the ability to present the
ubuntu users key. I've also cleaned up the language so it is clear how it works.