fix: fix cross-controller integration import (#980)
Fix cross-controller integration import support, allowing
the import to be matched up to the integration resource.

This is done by taking the controller name parsed out of the
offer string and looking for it in the offering_controller map
in the provider block.

Previous to this change, the provider would import and then
destroy/create the integration since it couldn't match it up.

Also add an import step to the cross-controller integration
test.