refactor(charmhub): removes the custom charmhub client for legacy deploy
Removes the custom charmhub client for legacy deploys and replaces it with a Charmhub facade call to
the Info method on 2.9 controllers. **NOTE** This change means that we do not support any Juju
version before 3.3 EXCEPT the 2.9 LTS version.

BREAKING CHANGE: Legacy deploy no longer works for Juju versions 3.0, 3.1 and 3.2.