fix(application): swallow unknown config error when unsetting config
It can happen a config value has been removed from the charm, if we still have it in the state,
terraform will try to unset it but Juju will return a `unknown config` error we need to swallow to
let the update through. This can introduce some delay when unsetting many configs because we do it
one by one.